Add check for html format to translations check
